requirements.txt
streamlit>=1.39
pandas>=2.0
numpy>=1.25
scikit-learn>=1.4
Pillow>=10.0
joblib>=1.4